Split mypy test for each release version
pytest was recently updated to a version that breaks the mypy tests. The
reason for that is because the new version of pytest uses a type
annotation that is only supported on python3.6 and later. Therefore
we cannot run all of the mypy tests using the latest version. Because
of that we are now splitting the mypy test for each release, making
sure the command is respecting the release dependencies freezes, which
include pytest